@@ -6,7 +6,7 @@ homepage: https://github.com/marica27/highlighter-coachmark
 maintainer: Marina Kuznetsova (@marica27)
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
