@@ -47,7 +47,7 @@ class CoachMark {
   CoachMark({this.bgColor = const Color(0xB2212121)});
 
   /// Global key to get an access for CoachMark's State
-  GlobalKey<_HighlighterCoachMarkState> globalKey;
+  GlobalKey<_HighlighterCoachMarkState>? globalKey;
 
   /// Background color
   Color bgColor;
@@ -59,10 +59,10 @@ class CoachMark {
   bool get isVisible => _isVisible;
 
   /// Called when CoachMark is closed
-  VoidCallback _onClose;
+  VoidCallback? _onClose;
 
   /// Contains OverlayEntry with CoachMark's Widget
-  OverlayEntry _overlayEntryBackground;
+  OverlayEntry? _overlayEntryBackground;
 
   /// Brings out CoachMark's widget with animation on the whole screen
   ///
@@ -84,12 +84,12 @@ class CoachMark {
   ///
   /// Callback [onClose] is called when CoachMark is closed
   void show({
-    @required BuildContext targetContext,
-    @required List<Widget> children,
-    @required Rect markRect,
+    required BuildContext targetContext,
+    required List<Widget> children,
+    required Rect markRect,
     BoxShape markShape = BoxShape.circle,
-    Duration duration,
-    VoidCallback onClose,
+    Duration? duration,
+    VoidCallback? onClose,
   }) async {
     // Prevent from showing multiple marks at the same time
     if (_isVisible) {
@@ -114,8 +114,8 @@ class CoachMark {
               ),
         );
 
-    OverlayState overlayState = Overlay.of(targetContext);
-    overlayState.insert(_overlayEntryBackground);
+    OverlayState overlayState = Overlay.of(targetContext)!;
+    overlayState.insert(_overlayEntryBackground!);
 
     if (duration != null) {
       await new Future.delayed(duration).then((_) => close());
@@ -126,11 +126,11 @@ class CoachMark {
   Future close() async {
     if (_isVisible) {
       await globalKey?.currentState?.reverse();
-      _overlayEntryBackground.remove();
+      _overlayEntryBackground!.remove();
 
       _isVisible = false;
       if (_onClose != null) {
-        _onClose();
+        _onClose!();
       }
     }
   }
@@ -140,12 +140,12 @@ class CoachMark {
 /// [markRect]
 class _HighlighterCoachMarkWidget extends StatefulWidget {
   _HighlighterCoachMarkWidget({
-    Key key,
-    @required this.markRect,
-    @required this.markShape,
-    @required this.children,
-    @required this.doClose,
-    @required this.bgColor,
+    Key? key,
+    required this.markRect,
+    required this.markShape,
+    required this.children,
+    required this.doClose,
+    required this.bgColor,
   }) : super(key: key);
 
   final Rect markRect;
@@ -160,9 +160,9 @@ class _HighlighterCoachMarkWidget extends StatefulWidget {
 
 class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _blurAnimation;
-  Animation<double> _opacityAnimation;
+  late AnimationController _controller;
+  late Animation<double> _blurAnimation;
+  late Animation<double> _opacityAnimation;
 
   //Does reverse animation, called when coachMark is closing.
   Future reverse() {
@@ -212,7 +212,7 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
 
     return AnimatedBuilder(
         animation: _controller,
-        builder: (BuildContext context, Widget child) {
+        builder: (BuildContext context, Widget? child) {
           return Stack(
             children: <Widget>[
               ClipPath(
@@ -267,14 +267,14 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
 /// overrides a special hitTest
 class _CoachMarkLayer extends Listener {
   const _CoachMarkLayer(
-      {Key key,
+      {Key? key,
       onPointerDown,
       onPointerMove,
       onPointerUp,
       onPointerCancel,
       behavior,
       this.markPosition,
-      Widget child})
+      Widget? child})
       : super(
             key: key,
             onPointerDown: onPointerDown,
@@ -283,7 +283,7 @@ class _CoachMarkLayer extends Listener {
             onPointerCancel: onPointerCancel,
             child: child);
 
-  final Rect markPosition;
+  final Rect? markPosition;
 
   @override
   RenderPointerListener createRenderObject(BuildContext context) {
@@ -319,9 +319,9 @@ class _RenderPointerListenerWithExceptRegion extends RenderPointerListener {
       onPointerMove,
       onPointerUp,
       onPointerCancel,
-      HitTestBehavior behavior,
+      required HitTestBehavior behavior,
       this.exceptRegion,
-      RenderBox child})
+      RenderBox? child})
       : super(
             onPointerDown: onPointerDown,
             onPointerMove: onPointerMove,
@@ -330,18 +330,18 @@ class _RenderPointerListenerWithExceptRegion extends RenderPointerListener {
             behavior: behavior,
             child: child);
 
-  final Rect exceptRegion;
+  final Rect? exceptRegion;
 
   @override
-  bool hitTest(HitTestResult result, {Offset position}) {
+  bool hitTest(HitTestResult result, {required Offset position}) {
     bool hitTarget = false;
-    if (exceptRegion.contains(position)) {
+    if (exceptRegion!.contains(position)) {
       result.add(new BoxHitTestEntry(this, position));
       return false;
     }
     if (size.contains(position)) {
       hitTarget =
-          hitTestChildren(result, position: position) || hitTestSelf(position);
+          hitTestChildren(result as BoxHitTestResult, position: position) || hitTestSelf(position);
       if (hitTarget || behavior == HitTestBehavior.translucent)
         result.add(new BoxHitTestEntry(this, position));
     }
@@ -367,15 +367,15 @@ class _CoachMarkClipper extends CustomClipper<Path> {
 ///This class makes edges of hole blurred.
 class _CoachMarkPainter extends CustomPainter {
   _CoachMarkPainter({
-    @required this.rect,
-    @required this.shadow,
+    required this.rect,
+    required this.shadow,
     this.clipper,
     this.coachMarkShape = BoxShape.circle,
   });
 
   final Rect rect;
   final BoxShadow shadow;
-  final _CoachMarkClipper clipper;
+  final _CoachMarkClipper? clipper;
   final BoxShape coachMarkShape;
 
   void paint(Canvas canvas, Size size) {
